% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Langevin.r
\name{plot.Langevin}
\alias{plot.Langevin}
\title{Plot estimated drift and diffusion coefficients}
\usage{
\method{plot}{Langevin}(x, pch = 20, ...)
}
\arguments{
\item{x}{an object of class "Langevin".}

\item{pch}{Either an integer specifying a symbol or a single character to be
used as the default in plotting points. See \link{points} for possible values
and their interpretation. Default is \code{pch = 20}.}

\item{...}{Arguments to be passed to methods, such as \code{\link{par}}.}
}
\description{
plot method for class "Langevin".
This method is only implemented for one-dimensional analysis for now.
}
\author{
Philip Rinn
}
