% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Langevin.r
\name{print.Langevin}
\alias{print.Langevin}
\title{Print estimated drift and diffusion coefficients}
\usage{
\method{print}{Langevin}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{an object of class "Langevin".}

\item{digits}{integer, used for number formatting with \code{\link{signif}()}.}

\item{...}{further arguments to be passed to or from other methods. They are
ignored in this function.}
}
\value{
The function \code{print.Langevin()} returns an overview of the
estimated drift and diffusion coefficients.
}
\description{
print method for class "Langevin".
}
\author{
Philip Rinn
}
