% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Langevin.r
\name{summary.Langevin}
\alias{summary.Langevin}
\title{Summarize estimated drift and diffusion coefficients}
\usage{
\method{summary}{Langevin}(object, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{an object of class "Langevin".}

\item{...}{arguments to be passed to or from other methods. They are
ignored in this function.}

\item{digits}{integer, used for number formatting with \code{\link{signif}()}.}
}
\value{
The function \code{summary.Langevin()} returns a sumamry of the
estimated drift and diffusion coefficients
}
\description{
summary method for class "Langevin".
}
\author{
Philip Rinn
}
